Download the SparkFun Eagle Library. This is the collection of all the components SparkFun designs with and therefore components and footprints that have been tested. Unzip and place the SparkFun.lbr file into the Eagle\lbr directory. If the above link does not work, google ?sparkfun eagle library? to get the latest collection.
Download the SparkFun Eagle keyboard shortcuts. Place this file in the Eagle\scr directory. If the above link does not work, google ?sparkfun eagle shortcut?.

Download the SparkFun CAM file. (Right-click, and choose Save Link As) Place this file in the Eagle\cam directory. This file is responsible for creating the gerber files for submission to a PCB fab house.